Transaction'un daha yalınlaşması için refactor.
