+++ package.json
@@ -32,7 +32,7 @@
     "yarn": "^1.17.3"
   },
   "scripts": {
-    "start": "set HTTPS=true && react-scripts start",
+    "start": "export HTTPS=true && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
